update default actor and reorganize files
