+++ simplemvi/src/commonMain/kotlin/com/arttttt/simplemvi/actor/ActorScope.kt
@@ -6,11 +6,13 @@ import kotlinx.coroutines.CoroutineScope
 interface ActorScope<in Intent : Any, State : Any, in SideEffect : Any> : CoroutineScope {
 
     val state: State
-    fun sideEffect(sideEffect: SideEffect)
 
     @MainThread
     fun intent(intent: Intent)
 
     @MainThread
     fun reduce(block: State.() -> State)
+
+    @MainThread
+    fun sideEffect(sideEffect: SideEffect)
 }
