+++ simplemvi/src/commonMain/kotlin/com/arttttt/simplemvi/utils/ActorDsl.common.kt
@@ -1,21 +1,9 @@
 package com.arttttt.simplemvi.utils
 
 import com.arttttt.simplemvi.actor.Actor
-import com.arttttt.simplemvi.actor.ActorScope
-import com.arttttt.simplemvi.actor.DefaultActor
 import com.arttttt.simplemvi.actor.dsl.ActorBuilder
 import com.arttttt.simplemvi.actor.dsl.DslActor
 import kotlin.coroutines.CoroutineContext
-
-fun <Intent : Any, State : Any, SideEffect : Any> defaultActor(
-    coroutineContext: CoroutineContext,
-    block: ActorScope<Intent, State, SideEffect>.(intent: Intent) -> Unit
-): Actor<Intent, State, SideEffect> {
-    return object : DefaultActor<Intent, State, SideEffect>(
-        coroutineContext = coroutineContext,
-        block = block,
-    ) {}
-}
 
 inline fun <Intent : Any, State : Any, SideEffect : Any> actorDsl(
     coroutineContext: CoroutineContext,
