+++ simplemvi/src/commonMain/kotlin/com/arttttt/simplemvi/utils/StoreExt.common.kt
@@ -1,8 +1,5 @@
 package com.arttttt.simplemvi.utils
 
-import com.arttttt.simplemvi.actor.Actor
-import com.arttttt.simplemvi.middleware.Middleware
-import com.arttttt.simplemvi.store.DefaultStore
 import com.arttttt.simplemvi.store.Store
 
 operator fun <Intent : Any> Store<Intent, *, *>.plus(intent: Intent) {
@@ -17,22 +14,3 @@ val <State : Any>Store<*, State, *>.state: State
     get() {
         return states.value
     }
-
-fun <Intent : Any, State : Any, SideEffect : Any> createStore(
-    initialize: Boolean = true,
-    initialState: State,
-    initialIntents: List<Intent> = emptyList(),
-    middlewares: List<Middleware<Intent, State, SideEffect>> = emptyList(),
-    actor: Actor<Intent, State, SideEffect>,
-): Store<Intent, State, SideEffect> {
-    return DefaultStore(
-        initialState = initialState,
-        initialIntents = initialIntents,
-        middlewares = middlewares,
-        actor = actor,
-    ).apply {
-        if (initialize) {
-            init()
-        }
-    }
-}
